I updated the readme with my favorite movies
Monster is an anime series adopted from the manga of Naoki Urusawa. The series explores upon the adventures of Dr. Tenma, a neurosurgeon, to kill a nameless monster.

The Cloverfield Trilogy are three movies, two of which are connected (1 and 2) and the third is a completely different genre switch altogether. The movie is amazing at making you feel suspense and fear of the unknown.

REC 1 and 2 are swedish movies that explore a deadly virus that would make the protagonists of the movie quarantined in buildings (in 1, it was a hotel. in 2, it was the airport) by the government in order to prevent the spread of the virus. It is perhaps the best zombie movie I've watched.

Dune 1 and 2 are movies adopted from the first book of the Dune novel. It is a very intriguing story following the life of Paul Atreides, a young adult forced to the barren desert planet of Arrakis.

Quentin Tarantino is an AMAZING director and nothing much is needed to be said other than watch him!